+++ relm4-macros/src/widgets/gen.rs
@@ -1,6 +1,6 @@
 use proc_macro2::{Span as Span2, TokenStream as TokenStream2};
 use quote::{quote, quote_spanned, ToTokens};
-use syn::{spanned::Spanned, Error, ExprCall, ExprField, Ident, Member, Path, Type};
+use syn::{spanned::Spanned, Error, ExprCall, ExprField, ExprPath, Ident, Member, Path, Type};
 
 use super::{Expr, Generics, Property, PropertyName, PropertyType, Tracker, Widget, WidgetFunc};
 
@@ -78,6 +78,46 @@ fn generate_tracker_from_expression(
     let bool_stream =
         quote_spanned! { expr_field.span() => model.changed( #model_name::#ident() ) };
     Ok(bool_stream)
+}
+
+fn component_ident(path: &ExprPath) -> TokenStream2 {
+    if path.path.segments.len() == 1 {
+        let ident = &path.path.segments.first().unwrap().ident;
+        quote_spanned! { path.span() => components.#ident.root_widget() }
+    } else {
+        path.to_token_stream()
+    }
+}
+
+fn component_tokens(expr: &Expr) -> TokenStream2 {
+    match expr {
+        Expr::Call(call) => {
+            if let Expr::Path(path) = &*call.func {
+                if let Some(segs) = path.path.segments.first() {
+                    if segs.ident == "Some" {
+                        if call.args.len() == 1 {
+                            if let Expr::Path(args_path) = call.args.first().unwrap() {
+                                let arg_tokens = component_ident(args_path);
+                                quote_spanned! { path.span() => Some(#arg_tokens) }
+                            } else {
+                                expr.to_token_stream()
+                            }
+                        } else {
+                            expr.to_token_stream()
+                        }
+                    } else {
+                        expr.to_token_stream()
+                    }
+                } else {
+                    expr.to_token_stream()
+                }
+            } else {
+                expr.to_token_stream()
+            }
+        }
+        Expr::Path(path) => component_ident(path),
+        _ => expr.to_token_stream(),
+    }
 }
 
 impl PropertyType {
@@ -169,7 +209,7 @@ impl PropertyType {
     fn component_tokens(&self) -> Option<TokenStream2> {
         match self {
             PropertyType::Widget(widget) => Some(widget.widget_assignment()),
-            PropertyType::Component(expr) => Some(quote_spanned! { expr.span() => #expr }),
+            PropertyType::Component(expr) => Some(component_tokens(expr)),
             _ => None,
         }
     }